Moved previous to_networkx test back to test_basics
